Make deprecation helper report total number of deprecations
Instead of listing all deprecations verbatim and returning
a non-zero exit code in case of deprecations just the total
number is reported.
The deprecations are only relevant if we want to upgrade to
Symfony 4 which as of now is not planned.
Furthermore, many deprecations can not easily be fixed because
they result from bundles that do not have an updated version, yet.